[rename] RoleSector -> PositionSector 파일명 변경
